requirements.txt
numpy
openai
python-dotenv
sounddevice
soundfile
playsound
pyttsx3
serial
pyserial
opencv-python
face_recognition
asyncio
mediapipe
pyaudio
pvrecorder
pygame
pvporcupine